+++ docs/source/releasenotes.rst
@@ -6,6 +6,7 @@ Upcoming release
 
 - Library **RPA.Email.ImapSmtp**: Return filepaths of saved attachments
 - Library **RPA.FileSystem**: Fix default argument handling (:issue:`170`)
+- Library **RPA.Word.Application**: Add option to control opening documents in ReadOnly mode (:issue:`171`)
 
 9.5.0
 -----